# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "DecisionCreate": {
        "description": "Record the human decision on an ``awaiting_review`` run.",
        "properties": {
          "decision": {
            "enum": [
              "approve",
              "override"
            ],
            "title": "Decision",
            "type": "string"
          },
          "note": {
            "default": "",
            "title": "Note",
            "type": "string"
          },
          "reviewer": {
            "default": "human",
            "title": "Reviewer",
            "type": "string"
          }
        },
        "required": [
          "decision"
        ],
        "title": "DecisionCreate",
        "type": "object"
      },
      "EvidenceEntry": {
        "description": "One auditable claim backing a score, a DDS field, or a narrative sentence.",
        "properties": {
          "artifact": {
            "additionalProperties": true,
            "title": "Artifact",
            "type": "object"
          },
          "cached": {
            "default": false,
            "title": "Cached",
            "type": "boolean"
          },
          "claim": {
            "title": "Claim",
            "type": "string"
          },
          "disclosure": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Disclosure"
          },
          "evidence_id": {
            "title": "Evidence Id",
            "type": "string"
          },
          "retrieved_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Retrieved At"
          },
          "schema_version": {
            "default": 1,
            "title": "Schema Version",
            "type": "integer"
          },
          "source": {
            "$ref": "#/components/schemas/EvidenceSource"
          },
          "synthetic": {
            "default": false,
            "title": "Synthetic",
            "type": "boolean"
          },
          "unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Unit"
          },
          "value": {
            "title": "Value"
          }
        },
        "required": [
          "evidence_id",
          "claim",
          "value",
          "source"
        ],
        "title": "EvidenceEntry",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunCreate": {
        "description": "Start one scenario run for a seed record or a named demo scenario.",
        "properties": {
          "model": {
            "anyOf": [
              {
                "enum": [
                  "scripted",
                  "bedrock"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "refresh": {
            "default": false,
            "title": "Refresh",
            "type": "boolean"
          },
          "scenario": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario"
          }
        },
        "title": "RunCreate",
        "type": "object"
      },
      "RunDetail": {
        "description": "Full dossier: trace, verification, assessment, review, and DDS metadata.",
        "properties": {
          "assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "dds": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/DdsMeta"
              },
              {
                "type": "null"
              }
            ]
          },
          "disclosures": {
            "items": {
              "type": "string"
            },
            "title": "Disclosures",
            "type": "array"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "expected_archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Archetype"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "pending_assessment": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "polygon_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon Id"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "review": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReviewOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "step_count": {
            "default": 0,
            "title": "Step Count",
            "type": "integer"
          },
          "steps": {
            "items": {
              "$ref": "#/components/schemas/TraceStep"
            },
            "title": "Steps",
            "type": "array"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          },
          "verification": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerificationReport"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "kind",
          "state",
          "model",
          "created_at"
        ],
        "title": "RunDetail",
        "type": "object"
      },
      "RunState": {
        "description": "Lifecycle of an assessment run (architecture run-state machine).",
        "enum": [
          "queued",
          "running",
          "needs_more_data",
          "awaiting_review",
          "complete",
          "failed"
        ],
        "title": "RunState",
        "type": "string"
      },
      "RunSummary": {
        "description": "One run row as it appears in list views and the cockpit table.",
        "properties": {
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "elapsed_seconds": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Elapsed Seconds"
          },
          "error": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Error"
          },
          "expected_archetype": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expected Archetype"
          },
          "finished_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Finished At"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "metrics": {
            "additionalProperties": true,
            "title": "Metrics",
            "type": "object"
          },
          "model": {
            "title": "Model",
            "type": "string"
          },
          "parent_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Run Id"
          },
          "polygon_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Polygon Id"
          },
          "record_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Record Id"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Score"
          },
          "started_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "state": {
            "$ref": "#/components/schemas/RunState"
          },
          "step_count": {
            "default": 0,
            "title": "Step Count",
            "type": "integer"
          },
          "summary": {
            "default": "",
            "title": "Summary",
            "type": "string"
          },
          "supplier_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Supplier Id"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Verdict"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "kind",
          "state",
          "model",
          "created_at"
        ],
        "title": "RunSummary",
        "type": "object"
      },
...
      "SupplierOut": {
        "properties": {
          "beneficial_owners": {
            "items": {
              "type": "string"
            },
            "title": "Beneficial Owners",
            "type": "array"
          },
          "certifications": {
            "items": {
              "type": "string"
            },
            "title": "Certifications",
            "type": "array"
          },
          "concession_area_ha": {
            "title": "Concession Area Ha",
            "type": "number"
          },
          "disclosure": {
            "default": "",
            "title": "Disclosure",
            "type": "string"
          },
          "group": {
            "title": "Group",
            "type": "string"
          },
          "hgu_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hgu Number"
          },
          "kabupaten": {
            "title": "Kabupaten",
            "type": "string"
          },
          "legal_name": {
            "title": "Legal Name",
            "type": "string"
          },
          "nib": {
            "title": "Nib",
            "type": "string"
          },
          "npwp": {
            "title": "Npwp",
            "type": "string"
          },
          "pbp_number": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Pbp Number"
          },
          "permit_status": {
            "title": "Permit Status",
            "type": "string"
          },
          "province": {
            "title": "Province",
            "type": "string"
          },
          "sanctions": {
            "items": {
              "type": "string"
            },
            "title": "Sanctions",
            "type": "array"
          },
          "supplier_id": {
            "title": "Supplier Id",
            "type": "string"
          },
          "synthetic": {
            "title": "Synthetic",
            "type": "boolean"
          },
          "trading_name": {
            "title": "Trading Name",
            "type": "string"
          }
        },
        "required": [
          "supplier_id",
          "legal_name",
          "trading_name",
          "group",
          "nib",
          "npwp",
          "permit_status",
          "concession_area_ha",
          "province",
          "kabupaten",
          "synthetic"
        ],
        "title": "SupplierOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
      "Verdict": {
        "description": "Outcome of the deterministic rubric for one supplier/parcel.",
        "enum": [
          "compliant",
          "high_risk",
          "ambiguous"
        ],
        "title": "Verdict",
        "type": "string"
      },
...
    }
  },
  "info": {
    "title": "TerraSentry API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/runs": {
      "get": {
        "operationId": "list_runs_runs_get",
        "parameters": [
          {
            "in": "query",
            "name": "kind",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Kind"
            }
          },
          {
            "in": "query",
            "name": "parent_run_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Parent Run Id"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RunSummary"
                  },
                  "title": "Response List Runs Runs Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Runs",
        "tags": [
          "runs"
        ]
      },
      "post": {
        "description": "Start the supervisor -> specialists -> verifier graph for one seed record.",
        "operationId": "create_run_runs_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RunCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}": {
      "get": {
        "operationId": "get_run_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/decision": {
      "post": {
        "description": "Approve or override an ambiguous run and release (or keep) its DDS.",
        "operationId": "decide_run_runs__run_id__decision_post",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DecisionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunSummary"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Decide Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/evidence": {
      "get": {
        "operationId": "get_evidence_runs__run_id__evidence_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/EvidenceEntry"
                  },
                  "title": "Response Get Evidence Runs  Run Id  Evidence Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Evidence",
        "tags": [
          "runs"
        ]
      }
    },
    "/runs/{run_id}/stream": {
      "get": {
        "description": "SSE stream of ``snapshot``, ``step``, ``state``, and ``done`` events.",
        "operationId": "stream_run_runs__run_id__stream_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Stream Run",
        "tags": [
          "runs"
        ]
      }
    },
    "/suppliers": {
      "get": {
        "operationId": "list_suppliers_suppliers_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SupplierOut"
                  },
                  "title": "Response List Suppliers Suppliers Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List Suppliers",
        "tags": [
          "suppliers"
        ]
      }
    },
    "/suppliers/{supplier_id}": {
      "get": {
        "operationId": "get_supplier_suppliers__supplier_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "supplier_id",
            "required": true,
            "schema": {
              "title": "Supplier Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SupplierDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Supplier",
        "tags": [
          "suppliers"
        ]
      }
    }
  }
}